[androsdk-402] Detekt and refactor tests
