+++ core/src/androidTest/java/org/hisp/dhis/android/core/user/UserDisabledMockIntegrationShould.kt
@@ -50,20 +50,6 @@ class UserDisabledMockIntegrationShould : BaseMockIntegrationTestMetadataEnqueab
         assertThat(d2.userModule().accountManager().getAccounts().size).isEqualTo(1)
     }
 
-    @Test
-    fun delete_database_when_user_disabled_on_rx_api_call_executor() {
-        dhis2MockServer.enqueueMockResponse(401, "user/user_disabled.json")
-        assertThat(d2.userModule().accountManager().getAccounts().size).isEqualTo(1)
-
-        try {
-            d2.eventModule().eventDownloader().blockingDownload()
-        } catch (e: Exception) {
-            val d2Error = e.cause as D2Error
-            assertThat(d2Error.errorCode()).isEqualTo(D2ErrorCode.USER_ACCOUNT_DISABLED)
-        }
-        assertThat(d2.userModule().accountManager().getAccounts().size).isEqualTo(0)
-    }
-
     private fun addDummyData() {
         d2.dataValueModule().dataValues().value(
             "20191021",
