+++ core/src/androidTest/java/org/hisp/dhis/android/core/user/UserDisabledRxAPIMockIntegrationShould.kt
@@ -1,0 +1,29 @@
+package org.hisp.dhis.android.core.user
+
+import com.google.common.truth.Truth.assertThat
+import org.hisp.dhis.android.core.maintenance.D2Error
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode
+import org.hisp.dhis.android.core.utils.integration.mock.BaseMockIntegrationTestEmptyEnqueable
+import org.hisp.dhis.android.core.utils.integration.mock.BaseMockIntegrationTestMetadataEnqueable
+import org.hisp.dhis.android.core.utils.runner.D2JunitRunner
+import org.junit.After
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(D2JunitRunner::class)
+class UserDisabledRxAPIMockIntegrationShould : BaseMockIntegrationTestEmptyEnqueable() {
+
+    //@Test
+    fun delete_database_when_user_disabled_on_rx_api_call_executor() {
+        dhis2MockServer.enqueueMockResponse(401, "user/user_disabled.json")
+        assertThat(d2.userModule().accountManager().getAccounts().size).isEqualTo(1)
+
+        try {
+            d2.eventModule().eventDownloader().blockingDownload()
+        } catch (e: Exception) {
+            val d2Error = e.cause as D2Error
+            assertThat(d2Error.errorCode()).isEqualTo(D2ErrorCode.USER_ACCOUNT_DISABLED)
+        }
+        assertThat(d2.userModule().accountManager().getAccounts().size).isEqualTo(0)
+    }
+}
